Refactor: 반환 타입 long 에서 int 로 수정 (#111)

+++ src/main/java/tavebalak/OTTify/review/repository/ReviewRepository.java
@@ -13,6 +13,7 @@ import tavebalak.OTTify.review.entity.Review;
 import tavebalak.OTTify.user.entity.User;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
+
     Slice<Review> findByUserIdOrderByCreatedAt(Long userId, Pageable pageable);
 
     boolean existsByProgramAndUser(Program program, User user);
@@ -36,7 +37,7 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
         @Param("genre") String genre, Pageable pageable);
 
     @Query("select count(r) from Review r where r.genre=:genre and r.program=:program")
-    long countByMyGenreName(@Param("genre") String genre, @Param("program") Program program);
+    int countByMyGenreName(@Param("genre") String genre, @Param("program") Program program);
 
     @Query("select sum(r.rating) from Review r where r.genre=:genre and r.program=:program")
     Double sumReviewRatingByGenreName(@Param("genre") String genreName,
